requirements.txt
Flask==2.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
click==8.0.3
joblib==1.1.0
nltk==3.6.5
numpy==1.21.4
pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
flask-cors==3.0.10
tensorflow==2.7.0